Update first-service
* generated from latest archetype
* add it as a sample instead
