--- docs/src/modules/java/examples/archetype/first-service/src/it/java/com/example/domain/CounterIntegrationTest.java
@@ -1,60 +1,0 @@
-package com.example.domain;
-
-import com.akkaserverless.javasdk.testkit.junit.AkkaServerlessTestkitResource;
-import com.example.CounterApi;
-import com.example.CounterServiceClient;
-import com.example.Main;
-import com.google.protobuf.Empty;
-import org.junit.ClassRule;
-import org.junit.Test;
-
-import static java.util.concurrent.TimeUnit.*;
-
-// Example of an integration test calling our service via the Akka Serverless proxy
-// Run all test classes ending with "IntegrationTest" using `mvn verify -Pit`
-public class CounterIntegrationTest {
-
-    /**
-     * The test kit starts both the service container and the Akka Serverless proxy.
-     */
-    @ClassRule
-    public static final AkkaServerlessTestkitResource testkit =
-            new AkkaServerlessTestkitResource(Main.createAkkaServerless());
-
-    /**
-     * Use the generated gRPC client to call the service through the Akka Serverless proxy.
-     */
-    private final CounterServiceClient client;
-
-    public CounterIntegrationTest() {
-        client = CounterServiceClient.create(testkit.getGrpcClientSettings(), testkit.getActorSystem());
-    }
-
-    @Test
-    public void increaseOnNonExistingEntity() throws Exception {
-        // TODO: set fields in command, and provide assertions to match replies
-        // client.increase(CounterApi.IncreaseValue.newBuilder().build())
-        //         .toCompletableFuture().get(2, SECONDS);
-    }
-
-    @Test
-    public void decreaseOnNonExistingEntity() throws Exception {
-        // TODO: set fields in command, and provide assertions to match replies
-        // client.decrease(CounterApi.DecreaseValue.newBuilder().build())
-        //         .toCompletableFuture().get(2, SECONDS);
-    }
-
-    @Test
-    public void resetOnNonExistingEntity() throws Exception {
-        // TODO: set fields in command, and provide assertions to match replies
-        // client.reset(CounterApi.ResetValue.newBuilder().build())
-        //         .toCompletableFuture().get(2, SECONDS);
-    }
-
-    @Test
-    public void getCurrentCounterOnNonExistingEntity() throws Exception {
-        // TODO: set fields in command, and provide assertions to match replies
-        // client.getCurrentCounter(CounterApi.GetCounter.newBuilder().build())
-        //         .toCompletableFuture().get(2, SECONDS);
-    }
-}
