+++ maven-java/akkaserverless-maven-archetype-value-entity/src/main/resources/archetype-resources/pom.xml
@@ -50,6 +50,9 @@
         <configuration>
           <source>${D}{jdk.target}</source>
           <target>${D}{jdk.target}</target>
+          <compilerArgs>
+            <arg>-Xlint:deprecation</arg>
+          </compilerArgs>
         </configuration>
       </plugin>
