+++ samples/first-service/src/main/java/com/example/Main.java
@@ -1,0 +1,30 @@
+/* This code was generated by Akka Serverless tooling.
+ * As long as this file exists it will not be re-generated.
+ * You are free to make changes to this file.
+ */
+
+package com.example;
+
+import com.akkaserverless.javasdk.AkkaServerless;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import com.example.domain.Counter;
+
+public final class Main {
+
+  private static final Logger LOG = LoggerFactory.getLogger(Main.class);
+
+  public static AkkaServerless createAkkaServerless() {
+    // The AkkaServerlessFactory automatically registers any generated Actions, Views or Entities,
+    // and is kept up-to-date with any changes in your protobuf definitions.
+    // If you prefer, you may remove this and manually register these components in a
+    // `new AkkaServerless()` instance.
+    return AkkaServerlessFactory.withComponents(
+      Counter::new);
+  }
+
+  public static void main(String[] args) throws Exception {
+    LOG.info("starting the Akka Serverless service");
+    createAkkaServerless().start();
+  }
+}
