+++ samples/first-service/src/main/proto/value-entities/counter_api.proto
