+++ samples/first-service/src/main/proto/value-entities/counter_domain.proto
