+++ samples/first-service/src/main/resources/logback.xml
@@ -11,7 +11,8 @@
                 <timestampFormat>yyyy-MM-dd'T'HH:mm:ss.SSSX</timestampFormat>
                 <timestampFormatTimezoneId>Etc/UTC</timestampFormatTimezoneId>
                 <appendLineSeparator>true</appendLineSeparator>
-                <jsonFormatter class="ch.qos.logback.contrib.jackson.JacksonJsonFormatter">
+                <jsonFormatter class="ch.qos.logback.contrib.jackson.JacksonJsonFormatter" >
+                    <!-- Don't use prettyPrint true in production, but can be useful for local development -->
                     <prettyPrint>false</prettyPrint>
                 </jsonFormatter>
             </layout>
