+++ samples/first-service/src/test/java/com/example/domain/CounterTest.java
@@ -1,0 +1,61 @@
+/* This code was generated by Akka Serverless tooling.
+ * As long as this file exists it will not be re-generated.
+ * You are free to make changes to this file.
+ */
+package com.example.domain;
+
+import com.akkaserverless.javasdk.testkit.ValueEntityResult;
+import com.akkaserverless.javasdk.valueentity.ValueEntity;
+import com.example.CounterApi;
+import com.google.protobuf.Empty;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.NoSuchElementException;
+import org.junit.Test;
+import scala.jdk.javaapi.CollectionConverters;
+
+import static org.junit.Assert.*;
+
+public class CounterTest {
+
+  @Test
+  public void exampleTest() {
+    CounterTestKit testKit = CounterTestKit.of(Counter::new);
+    // use the testkit to execute a command
+    // of events emitted, or a final updated state:
+    // ValueEntityResult<SomeResponse> result = testKit.someOperation(SomeRequest);
+    // verify the response
+    // SomeResponse actualResponse = result.getReply();
+    // assertEquals(expectedResponse, actualResponse);
+    // verify the final state after the command
+    // assertEquals(expectedState, testKit.getState());
+  }
+
+  @Test
+  public void increaseTest() {
+    CounterTestKit testKit = CounterTestKit.of(Counter::new);
+    // ValueEntityResult<Empty> result = testKit.increase(IncreaseValue.newBuilder()...build());
+  }
+
+
+  @Test
+  public void decreaseTest() {
+    CounterTestKit testKit = CounterTestKit.of(Counter::new);
+    // ValueEntityResult<Empty> result = testKit.decrease(DecreaseValue.newBuilder()...build());
+  }
+
+
+  @Test
+  public void resetTest() {
+    CounterTestKit testKit = CounterTestKit.of(Counter::new);
+    // ValueEntityResult<Empty> result = testKit.reset(ResetValue.newBuilder()...build());
+  }
+
+
+  @Test
+  public void getCurrentCounterTest() {
+    CounterTestKit testKit = CounterTestKit.of(Counter::new);
+    // ValueEntityResult<CurrentCounter> result = testKit.getCurrentCounter(GetCounter.newBuilder()...build());
+  }
+
+}
